Fix for jQuery preset test timing out
`''` is falsey, so when the CLI did

```javascript
var configPath = path.resolve(process.cwd(), program.config ||
'.jscs.json');
```

it was using the main `.jscs.json` file from the root, then adding the
jquery rules to that. That seems like it would create strange
side-effects and wasn't intended, so I switched it to a value that would
actually cause the `doesConfigExist` value to be what was expected.

Incidentally, JSCS now finds 2 errors in `test/data/cli.js` so I changed
the comparison to be a little more broad so it'll be less fragile in the
future.