Update requireMultipleVarDecl rule to support one var per declaration
requireMultipleVarDecl only supported one var per function scope and
consecutive var checks. It was not possible to check that each declaration
statement had it's own var terminated by a semicolon such as in the npm
javascript guide. See
https://github.com/Seravo/js-winning-style/blob/master/README.md#variable-declarations

This means the original requireMultipleVarDecl option of boolean true for
consectutive var check is now changed to a string type aptly named
'nonconsecutive', and a new option of 'onedecl' is introduced for the new
check.