Config: Add "ignore-missing-config" option
Currently, jscs exits with a non-zero status code when it can't find a config
file. This is great for certain use cases, e.g. as a reminder for when you have
forgotten to add a .jscsrc file to a project directory. But for other
situations, e.g. when you have configured your editor to send any opened "*.js"
file to jscs, and your editor warns you when the program exits with a non-zero
status code, then it can be quite annoying to try and edit files from projects
which do not use jscs, as you will be bombarded with "No configuration found."
messages. One workaround is to "echo '{}' >> ~/.jscsrc", but it's not a very
good solution. So this option allows you to no-op your way through the program
if a config file is not found.