Reporters: make inline and inlinesingle machine readable
Changes the formatting of the reporters `inline` and `inlinesingle`
from

	file.js: line 7, col 7, Unexpected token {

to

	file.js:7:7: Unexpected token {

which is much like the output from other compilers and preprocessors.

The benefit is integration with existing tools that parse similar
output.  For example Emacs allows navigation by these “hyperlinks”
through the rules in `compilation-error-regexp-alist`.  Other
editors, like acme, lets you right-click on such tokens to jump to
the file, line, and column.

Closes gh-1401