Reporters: add new machine readable unix-style reporter
Similar to `inline` and `inlinesingle`, but instead of outputting

	file.js: line 7, col 7, Unexpected token {

it outputs

	file.js:7:7: Unexpected token {

which is much like the output from other compilers and preprocessors.

The benefit is integration with existing tools that parse similar
output.  For example Emacs allows navigation by these “hyperlinks”
through the rules in `compilation-error-regexp-alist`.  Other editors
such as acme, lets you right-click on such tokens to jump to the
file, line, and column.

Closes gh-1403