requireBlocksOnNewline: Add object option to handle comments
Add object option `{ includeComments: true }` and
`{ includeComments: true, minLines: 1} `

that allows the rule to turn on or off the inclusion of comments as part
of the block being checked using exisitng rules.

A new config option is introduced to maintain compatability with existing
code like the crockford preset which ignores comments.

Fixes #1417
Closes gh-1620