requireCamelCaseOrUpperCaseIdentifiers: add prefix/suffix options
Some projects may want to use this rule but require that values that
hold measurements carry a suffix specifying the units of measurement;
e.g. `ambientTemperature_dCel` or `centerFrequency_kHz`.  The
interpretation of units can be misleading when conjoined with the
identifier without a separating underscore.

Similarly some coding guidelines might want optional parameters to start
with `opt_` while still checking conformance of the remainder of the
name.

This commit adds two options, `allowedPrefixes` and `allowedSuffixes`,
both of which recognize an array of strings or RegExps that match units
that are allowed to appear as an underscore-separated prefix (suffix) of
an otherwise acceptable identifier.

Fixes #2029
Closes gh-2028