requireCamelCaseOrUpperCaseIdentifiers: add exception options
Some coding guidelines might want optional parameters to start with
`opt_` while still checking conformance of the remainder of the name, or
have specific identifiers like `var_args` that are exceptions.

Other projects may want to use this rule but require that values that
hold measurements carry a suffix specifying the units of measurement;
e.g. `ambientTemperature_dCel` or `centerFrequency_kHz`.  The
interpretation of units can be misleading when conjoined with the
identifier without a separating underscore.

This commit adds options `allExcept`, `allowedPrefixes`, and
`allowedSuffixes`.  All three recognize an array of strings, RegExps, or
ESTree RegExpLiteral objects that identify strings or string sequences.
Identifiers that match strings/patterns in `allExcept` are excluded from
validation by the rule.  Identifiers that start (end) with a match from
`allowedPrefixes` (`allowedSuffixes`) are checked after the match is
removed from the identifier.

Fixes #1900
Fixes #2029